requirements.txt
requests==2.31.0
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.0
pydantic==1.10.13
streamlit==1.31.1
plotly==5.18.0
beautifulsoup4==4.12.2
pdfkit==1.0.0
pillow==10.1.0
reportlab==3.6.13